Continue with visuals
Right pane with buttons is now with determined dimensions, so only the
left pane flexes - otherwise it was difficult to set the buttons of
computer's choice on the proper positions below. Definitely won't
work for mobiles at the moment.
Small changes regarding graphics:
- computer's choice is displayed on a smaller background, it's
position is just below the corresponding button of player's choice
- latest score blinks with orange color
- final result moved to the right side: work in progress